django_refresh/requirements/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.5
Django==2.1.7
django-rest-knox==3.6.0
djangorestframework==3.9.1
pkg-resources==0.0.0
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
